removed debug folder
